move availalbe bulk actions inside bulk aciton compoonent
